Extract cursor move operations to a separate file
